Eliminando arena_transitional.ts
Este archivo ya cumplió su objetivo y ya no es necesario tenerlo por
separado.
