+++ frontend/www/js/omegaup/arena/admin_arena.ts
@@ -7,8 +7,7 @@ import arena_Runs from '../components/arena/Runs.vue';
 import * as ui from '../ui';
 import * as time from '../time';
 
-import { Arena } from './arena';
-import { runsStore } from './arena_transitional';
+import { Arena, runsStore } from './arena';
 
 export default class ArenaAdmin {
   arena: Arena;
